\name{AddConMat}
\alias{AddConMat}
\title{INTERNAL
Adds the overlap edges to a mask database}
\usage{
  AddConMat(db, cmat, ids)
}
\arguments{
  \item{db}{the mask database object}

  \item{cmat}{a matrix of 0/1 values giving the locations
  of edges between masks}

  \item{ids}{a vector giving the ids of the masks in cmat
  (in order)}
}
\value{
  NULL
}
\description{
  INTERNAL Adds the overlap edges to a mask database
}

