\name{AddMasks}
\alias{AddMasks}
\title{Generate masks according to the given method and add them to the database}
\usage{
  AddMasks(db, calexp, method)
}
\arguments{
  \item{db}{the mask database object}

  \item{calexp}{the calcium experiment data object}

  \item{method}{what method should be used to generate
  masks to add}
}
\description{
  Generate masks according to the given method and add them
  to the database
}

