\name{AssignContiguous}
\alias{AssignContiguous}
\title{Assigns the non-zero pixels of 'region' to one contiguous (cardinal directions) regions}
\usage{
  AssignContiguous(region)
}
\arguments{
  \item{region}{a matrix with 1 in the regions to be
  assigned and 0 elsewhere}
}
\value{
  a matrix with unique integers in the pixels of each
  region
}
\description{
  Assigns the non-zero pixels of 'region' to one contiguous
  (cardinal directions) regions
}

