\name{ClustDistance}
\alias{ClustDistance}
\title{Return the corrected Rand index for the two given clusterings}
\usage{
  ClustDistance(clust, clust2)
}
\arguments{
  \item{clust}{a vector giving the first clustering}

  \item{clust2}{a vector giving the second clustering}
}
\value{
  the corrected Rand index value
}
\description{
  Return the corrected Rand index for the two given
  clusterings
}

