\name{ClusterCorrelation}
\alias{ClusterCorrelation}
\title{See correlation with clusters}
\usage{
  ClusterCorrelation(calexp, clusters)
}
\arguments{
  \item{calexp}{the calexp object}

  \item{clusters}{the cluster object as returned from
  ClusterCells}
}
\description{
  See correlation with clusters
}

