\name{CreateDbController}
\alias{CreateDbController}
\title{INTERNAL
Creates an empty database controller}
\usage{
  CreateDbController()
}
\value{
  A list \item{db.directory}{the directory holding SQLite
  databases} \item{data.directory}{the directory holding
  data associated with each database.  each of these should
  have a \$data element} \item{helper.directory}{the
  directory in which to place helper files generated by the
  GUI} \item{classifier.direcoty}{the directory that
  contains the classifiers used in segmentation}
  \item{expdf}{a data.frame in which to put information
  about each experiment, currently empty}
}
\description{
  INTERNAL Creates an empty database controller
}
\details{
  A database controller holds information about the
  directories where the databases, data, classifiers, and
  helper files are stored. \cr
}

