\name{DbSetup}
\alias{DbSetup}
\title{INTERNAL
Creates an empty mask database with the appropriate tables}
\usage{
  DbSetup(db)
}
\arguments{
  \item{db}{the database object for which to create the
  mask tables}
}
\value{
  NULL
}
\description{
  INTERNAL Creates an empty mask database with the
  appropriate tables
}

