\name{GetPhase}
\alias{GetPhase}
\title{Gets the phase of a given frequency band of a vector using the FFT}
\usage{
  GetPhase(vec, low, high, dt = 1/1000)
}
\arguments{
  \item{vec}{vector of data}

  \item{low}{lower bound of frequency band}

  \item{high}{upper bound of frequency band}

  \item{dt}{deltaT, or 1/sampling rate}
}
\description{
  Gets the phase of a given frequency band of a vector
  using the FFT
}

