\name{GKernel}
\alias{GKernel}
\title{Returns a Gaussian kernel}
\usage{
  GKernel(kdim, sigma)
}
\arguments{
  \item{kdim}{the dimension of the (square) kernel to
  generate}

  \item{sigma}{the standard deviation of the gaussian}
}
\value{
  a matrix giving the Gaussian kernel
}
\description{
  Returns a Gaussian kernel
}

