\name{Image}
\alias{Image}
\title{Plots an image of the given matrix with the origin in the upper left}
\usage{
  Image(img, col = grey(seq(0, 1, 0.001)), ...)
}
\arguments{
  \item{img}{the image matrix to plot}

  \item{col}{a list of colors to use for plotting, defaults
  to grey}

  \item{...}{additional graphing parameters}
}
\value{
  NULL
}
\description{
  Plots an image of the given matrix with the origin in the
  upper left
}

