\name{InitiateMaskClassifier}
\alias{InitiateMaskClassifier}
\title{Creates a mask classifier based on the given training data}
\usage{
  InitiateMaskClassifier(trainingdata)
}
\arguments{
  \item{trainingdata}{the training data, as returned by
  PullAllData or PullData}
}
\value{
  a mask classifier
}
\description{
  Creates a mask classifier based on the given training
  data
}

