\name{InvertMask}
\alias{InvertMask}
\title{INTERNAL
Inverts a mask matrix so that the mask region is turned to backgroun
and vice versa}
\usage{
  InvertMask(mask)
}
\arguments{
  \item{mask}{the mask matrix to invert, with NA in the
  background}
}
\value{
  a matrix with the inverted mask
}
\description{
  INTERNAL Inverts a mask matrix so that the mask region is
  turned to backgroun and vice versa
}

