\name{LoGKernel}
\alias{LoGKernel}
\title{Returns a Laplacian of Gaussian kernel}
\usage{
  LoGKernel(kdim, sigma)
}
\arguments{
  \item{kdim}{the dimension of the (square) kernel to
  generate}

  \item{sigma}{the standard deviation of the gaussian
  smoother}
}
\value{
  a matrix giving the LoG kernel
}
\description{
  Returns a Laplacian of Gaussian kernel
}

