\name{MaskDbSetup}
\alias{MaskDbSetup}
\title{INTERNAL
Creates an empty mask database with the appropriate tables}
\usage{
  MaskDbSetup(db, calexp, tag, dt = NA, description = NA,
    datafile = NA)
}
\arguments{
  \item{db}{the database object for which to create the
  mask tables}
}
\value{
  NULL
}
\description{
  INTERNAL Creates an empty mask database with the
  appropriate tables
}

