\name{MatrixToSparseMasks}
\alias{MatrixToSparseMasks}
\title{Converts a matrix with unique positive integers on each mask into a sparse mask list}
\usage{
  MatrixToSparseMasks(maskmat)
}
\arguments{
  \item{maskmat}{the matrix with unique integers on each
  mask}

  \item{return}{a sparsemasks object with a list of masks
  in the $masks element and the dimensions of the original
  image in the $dims element}
}
\description{
  Converts a matrix with unique positive integers on each
  mask into a sparse mask list
}

