\name{MultiTaperSpectrum}
\alias{MultiTaperSpectrum}
\title{Uses multi-taper methods to etimate a spectrum for the given vector}
\usage{
  MultiTaperSpectrum(vec, dt = 0.1247232, dif = T)
}
\arguments{
  \item{vec}{the vector of data}

  \item{dt}{deltaT, or 1/sampling rate}

  \item{dif}{boolean, should the vector be differenced
  before estimating the spectrum}
}
\description{
  Uses multi-taper methods to etimate a spectrum for the
  given vector
}

