% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_check.R
\name{num_check}
\alias{num_check}
\title{Check variable for numeric data type}
\usage{
num_check(df, var)
}
\arguments{
\item{df}{Data frame}

\item{var}{Variable}
}
\value{
Message (if variable is not of a numeric data type)
}
\description{
This function checks a given variable for a numeric data type
}
