\name{PhaseDist}
\alias{PhaseDist}
\title{Compute circular phase distance between two phases}
\usage{
  PhaseDist(phase1, phase2)
}
\arguments{
  \item{phase1}{a phase value in [-pi,pi]}

  \item{phase2}{a phase value in [-pi,pi]}
}
\value{
  the circular distance between the two phases
}
\description{
  Compute circular phase distance between two phases
}

