\name{PlotMaskSetByID}
\alias{PlotMaskSetByID}
\title{Plots masks specified by the given ids}
\usage{
  PlotMaskSetByID(db, ids, rgb = NULL)
}
\arguments{
  \item{db}{the database}

  \item{ids}{a vector giving the ids of the masks to plot}

  \item{rgb}{a vector of length 3 giving the color to plot
  the masks}
}
\value{
  NULL
}
\description{
  Plots masks specified by the given ids
}

