\name{PlotSegmentation}
\alias{PlotSegmentation}
\title{Plots the segmentation of a particular class, as stored in a mask database}
\usage{
  PlotSegmentation(db, classid, rgb = NULL)
}
\arguments{
  \item{db}{the database}

  \item{classid}{the class to plot}

  \item{rgb}{a vector of length 3 giving the color to plot
  the masks}
}
\value{
  NULL
}
\description{
  Plots the segmentation of a particular class, as stored
  in a mask database
}

