\name{PullAllData}
\alias{PullAllData}
\title{Checks each database in the given directory and pulls any
labeled data}
\usage{
  PullAllData(directory)
}
\arguments{
  \item{directory}{the directory to pull data from}
}
\value{
  a data.frame with the extracted data
}
\description{
  Checks each database in the given directory and pulls any
  labeled data
}

