% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rci.R
\name{rci}
\alias{rci}
\title{Computes Reliable Change Index (RCI)}
\usage{
rci(df, pretest, posttest, method = "JT")
}
\arguments{
\item{df}{Data Frame}

\item{pretest}{Column Name of Pre-test scores}

\item{posttest}{Column Name of Post-test score}

\item{method}{RCI Formula to use}
}
\value{
Original Data Frame with newly added column containing RCI scores
}
\description{
This function computes the Reliable Change Index (RCI) for a given psychometric instrument across 2 measurement points
}
\examples{
# Data frame containing of Pre-test and Post-test columns
set.seed(5)
df <- data.frame(ID = c("AAA", "BBB", "CCC", "DDD", "EEE"), W1 = rnorm(5), W2 = rnorm(5))
# Input the data frame, Pre-test column, and Post-test columns as arguments, respectively
rci(df, "W1", "W2")

}
