\name{RemoveMask}
\alias{RemoveMask}
\title{Remove a mask from a mask database}
\usage{
  RemoveMask(db, maskid)
}
\arguments{
  \item{db}{a database connection}

  \item{maskid}{the ID of the mask to remove}
}
\value{
  NULL
}
\description{
  Remove a mask from a mask database
}
\details{
  Removes a mask, as well as any associated features and
  edges
}

