\name{RotateImg}
\alias{RotateImg}
\title{INTERNAL
Rotates an image by a given number of integer rows and columns}
\usage{
  RotateImg(mat, x, y)
}
\arguments{
  \item{mat}{the matrix to rotate}

  \item{x}{the number of columns to rotate}

  \item{y}{the number of rows to rotate}
}
\value{
  the rotated matrix
}
\description{
  INTERNAL Rotates an image by a given number of integer
  rows and columns
}

