\name{ShiftVector}
\alias{ShiftVector}
\title{INTERNAL
Shifts a vector by the specified amount using FFT}
\usage{
  ShiftVector(vec, amt)
}
\arguments{
  \item{vec}{the vector to shift}

  \item{amt}{the amount to shift}
}
\value{
  the circularly shifted vector
}
\description{
  INTERNAL Shifts a vector by the specified amount using
  FFT
}

