\name{ViewCI}
\alias{ViewCI}
\title{Opens the GUI viewer to manipulate the segmentation process.}
\usage{
  ViewCI(dbController = NULL)
}
\arguments{
  \item{dbController}{if specified, the viewer opens with
  the given dbController (looking in the directories stored
  in that object)}
}
\value{
  NULL
}
\description{
  Opens the GUI viewer to manipulate the segmentation
  process.
}

