\name{summary.MaskDb}
\alias{summary.MaskDb}
\title{Summary function to quickly see the details/statistics of a mask database object}
\usage{
  summary.MaskDb(db)
}
\arguments{
  \item{db}{the mask database option}
}
\value{
  NULL
}
\description{
  Summary function to quickly see the details/statistics of
  a mask database object
}

