\name{ViewCI}
\alias{ViewCI}
\title{#INTERNAL
#Creates an empty database controller
#
#}
\usage{
  ViewCI(db, calexp)
}
\arguments{
  \item{db}{if specified, the viewer opens with the given
  dbController (looking in the directories stored in that
  object)}
}
\value{
  A list #\item{db.directory}{the directory holding SQLite
  databases} #\item{data.directory}{the directory holding
  data associated with each database. # each of these
  should have a \$data element}
  #\item{helper.directory}{the directory in which to place
  helper files generated by the GUI}
  #\item{classifier.direcoty}{the directory that contains
  the classifiers used in segmentation} #\item{expdf}{a
  data.frame in which to put information about each
  experiment, currently empty} #INTERNAL #Create an object
  to store information about the currently selected
  experiment. # #

  A list with fields to store information about the
  experiment (initially empty) #\item{name}{the experiment
  name} #\item{data}{if loaded, the data object for this
  object} #\item{db}{the database connection for this
  experiment} #\item{nmasks}{the number of candidate masks
  in the database for this experiment} #\item{features}{the
  tags of the features that exist in this database}
  #\item{sources}{the tage for the mask sources present in
  this database} #\item{selmat}{a matrix where the first
  column is the ID of the mask and the second column gives
  the annotation for the mask} #\item{sms}{the list of
  sparse masks for the experiment retreived from the
  database with GetMasks()} #\item{mimg1}{the matrix giving
  the mean image for channel 1} #\item{mimg2}{the matrix
  giving the mean image for channel 2} #\item{nx}{the
  number of columns in the images for this experiment}
  #\item{ny}{the number of rows in the images} Opens the
  GUI viewer to manipulate the segmentation process.

  NULL
}
\description{
  #INTERNAL #Creates an empty database controller # #
}
\details{
  A database controller holds information about the
  directories #where the databases, data, classifiers, and
  helper files are stored. \cr # #
}

